Move protocol detector injection outside of sync block
Fixes ViaVersion/ViaFabric#302
