+++ viafabric-mc1204/src/main/resources/mixins.viafabric1204.pipeline.json
@@ -6,7 +6,7 @@
     "MixinClientConnection"
   ],
   "client": [
-    "client.MixinClientConnection"
+    "client.MixinClientConnectionThread"
   ],
   "injectors": {
     "defaultRequire": 1
