+++ viafabric-mc1206/src/main/resources/mixins.viafabric1206.pipeline.json
@@ -6,7 +6,7 @@
     "MixinClientConnection"
   ],
   "client": [
-    "client.MixinClientConnection"
+    "client.MixinClientConnectionThread"
   ],
   "injectors": {
     "defaultRequire": 1
