+++ viafabric-mc1211/src/main/java/com/viaversion/fabric/mc1211/mixin/pipeline/client/MixinClientConnectionThread.java
@@ -15,25 +15,24 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package com.viaversion.fabric.mc1206.mixin.pipeline.client;
+package com.viaversion.fabric.mc1211.mixin.pipeline.client;
 
-import com.viaversion.fabric.mc1206.ViaFabric;
-import com.viaversion.fabric.mc1206.service.ProtocolAutoDetector;
-import io.netty.channel.ChannelFuture;
-import net.minecraft.network.ClientConnection;
+import com.llamalad7.mixinextras.sugar.Local;
+import com.viaversion.fabric.mc1211.ViaFabric;
+import com.viaversion.fabric.mc1211.service.ProtocolAutoDetector;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
-import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
+import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 
 import java.net.InetSocketAddress;
 import java.util.concurrent.TimeUnit;
 import java.util.logging.Level;
 
-@Mixin(ClientConnection.class)
-public class MixinClientConnection {
-    @Inject(method = "connect(Ljava/net/InetSocketAddress;ZLnet/minecraft/network/ClientConnection;)Lio/netty/channel/ChannelFuture;", at = @At("HEAD"))
-    private static void onConnect(InetSocketAddress address, boolean useEpoll, ClientConnection connection, CallbackInfoReturnable<ChannelFuture> cir) {
+@Mixin(targets = "net/minecraft/client/gui/screen/multiplayer/ConnectScreen$1")
+public class MixinClientConnectionThread {
+    @Inject(method = "run", at = @At(value = "INVOKE_ASSIGN", args = "fuzz=2", target = "Ljava/util/Optional;get()Ljava/lang/Object;"))
+    private static void onConnect(CallbackInfo ci, @Local InetSocketAddress address) {
         try {
             if (!ViaFabric.config.isClientSideEnabled()) return;
             ProtocolAutoDetector.detectVersion(address).get(10, TimeUnit.SECONDS);
