+++ viafabric-mc1211/src/main/resources/mixins.viafabric1211.pipeline.json
@@ -6,7 +6,7 @@
     "MixinClientConnection"
   ],
   "client": [
-    "client.MixinClientConnection"
+    "client.MixinClientConnectionThread"
   ],
   "injectors": {
     "defaultRequire": 1
